Step 42
With form submissions, it is useful, and good practice, to provide each submittable element with a name attribute. This attribute is used to identify the element in the form submission.

Except for the two radio inputs (which you have already named), give each submittable element a unique name attribute of your choosing.