Step 13
Align your .header text in the center. Make the text uppercase using the text-transform property with uppercase as the value.

Give it a padding of 32px on all sides. Set the background to #0a0a23 and the text to #fff as the color values.

Add a border-bottom with 4px solid #fdb347 as the value.